Trend line rendering moved to plugin from core, plus some cruft cleanup in TimeSeriesPlotter.render()
